Add support for corrupted/empty entries
When loading phone numbers, we use the #attributes method of a Nori-parsed
node to access @label and @rel metadata. However, if the XML node is empty,
Nori will not parse it and instead return a hash, which lacks an #attributes
method and crashes the parser with a undefined method error.

To avoid these malformed entries from crashing the parser, this commit adds a
quick #respond_to? check before processing as a way of validating the phone
data. Entries which do not respond to #attributes are assumed to be corrupted
or empty and are discarded. Other valid phone numbers from within the contact
entry are added as usual.